+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [5.12.0](https://github.com/forcedotcom/source-deploy-retrieve/compare/v5.11.0...v5.12.0) (2022-02-22)
+
+### Features
+
+- **metadata registry:** adding flowtest type to metadata registry ([#580](https://github.com/forcedotcom/source-deploy-retrieve/issues/580)) ([40431f6](https://github.com/forcedotcom/source-deploy-retrieve/commit/40431f60eb29d02e33b5fc4f5f2066271b06c630))
+
+## [5.11.0](https://github.com/forcedotcom/source-deploy-retrieve/compare/v5.10.0...v5.11.0) (2022-02-22)
+
+### Features
+
+- cache fileResponses on deployResult ([#573](https://github.com/forcedotcom/source-deploy-retrieve/issues/573)) ([b38185b](https://github.com/forcedotcom/source-deploy-retrieve/commit/b38185b7cd73b05ef1e1940b878b09eb2d198101))
+
 ## [5.10.0](https://github.com/forcedotcom/source-deploy-retrieve/compare/v5.9.11...v5.10.0) (2022-02-17)
 
 ### Features